requirements.txt
# --- UI ---
gradio==4.44.1
gradio-client==1.3.0

# --- Búsqueda / ML ---
sentence-transformers>=2.7,<3
faiss-cpu>=1.8.0,<2

# --- Dependencias base ---
numpy>=1.26,<3
torch>=2.2,<3